% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_dist.R
\name{calc_p_value_safer}
\alias{calc_p_value_safer}
\title{Survival function of truncated normal distribution.
Log-sum-exp operations are used to avoid underflows in the upper tail probability of
a truncated normal distribution}
\usage{
calc_p_value_safer(truncation, vTy, nu_norm, sig, mu = 0, two_sided = TRUE)
}
\description{
Let \eqn{X} be a normal random variable with mean \code{mu} and standard deviation \code{sig*nu_norm}.
Truncating \eqn{X} to the set \eqn{df} is equivalent to conditioning on \eqn{{X \in df}}.
So this function returns \eqn{P(|X| \ge v^{T}y | X \in df)} .
}
