% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{check_segment_in_model}
\alias{check_segment_in_model}
\title{This model takes in
\enumerate{
\item @new_cluster new cluster assignment
\item @old_segment a list segment (represented by the positions of the nodes),
and checks if the list of segments of interest is in the new graph
(assume same ordering of nodes)
}}
\usage{
check_segment_in_model(old_segment, new_cluster)
}
\description{
This model takes in
\enumerate{
\item @new_cluster new cluster assignment
\item @old_segment a list segment (represented by the positions of the nodes),
and checks if the list of segments of interest is in the new graph
(assume same ordering of nodes)
}
}
