% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_p_val.R
\name{test_gfl_approx}
\alias{test_gfl_approx}
\title{Monte Carlo significance test for any clustering method}
\usage{
test_gfl_approx(
  y = y_vec,
  Dmat = Dmat,
  v = contrast,
  K = K_init,
  sigma = sigma,
  segment_list,
  two_sided = TRUE,
  ndraws = 1000,
  stop_criteria = "K",
  random_seed = 2021
)
}
\description{
Monte Carlo significance test for any clustering method
}
