% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_funcs.R
\name{TNSurv}
\alias{TNSurv}
\title{Survival function of truncated normal distribution.}
\usage{
TNSurv(truncation, vTy, nu_norm, sig, mu = 0)
}
\arguments{
\item{truncation}{A data frame of truncation with 3 columns: min_mean, max_mean, and contained.
Each row corresponds to a truncation interval with lower and upper limits specified by min_mean and max_mean,
respectively. In addition, contained is binary-valued and indicates whether this interval is included
in the final truncation set.}

\item{vTy}{Numeric; the value to evaluate the survival function on}

\item{nu_norm}{Numeric; part 1 of standard deviation}

\item{sig}{Numeric; part 2 of standard deviation}

\item{mu}{Numeric; mean of untruncated distribution, default to 0}
}
\description{
Survival function of truncated normal distribution.
}
\details{
Let \eqn{X} be a normal random variable with mean \code{mu} and standard deviation \code{sig*nu_norm}.
This function returns \eqn{P(X \ge v^{T}y | X \in \code{truncation)}}, where \code{truncation} is a subset of the real line.
Log-sum-exp operations are used to avoid underflows in the upper tail probability of the truncated normal distribution.

Input:
}
\examples{
truncation <- data.frame(matrix(c(-2,1,0,2),byrow = T,ncol=2))
colnames(truncation) <- c("min_mean", "max_mean")
truncation$contained <- 1
TNSurv(truncation,vTy=0, nu_norm=1, sig=1)
}
\keyword{internal}
