% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{check_segment_in_model}
\alias{check_segment_in_model}
\title{check_segment_in_model; util function}
\usage{
check_segment_in_model(old_segment, new_cluster)
}
\description{
check_segment_in_model; util function
}
\details{
This model takes in
1. new_cluster new cluster assignment
2. old_segment a list segment (represented by the positions of the nodes),
and checks if the list of segments of interest is in the new graph
(assume same ordering of nodes)
}
\keyword{internal}
