% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_funcs.R
\name{compute_CI}
\alias{compute_CI}
\title{compute selective confidence intervals}
\usage{
compute_CI(vTy, vTv, sigma, truncation, alpha)
}
\arguments{
\item{vTy}{y: the observed response vector}

\item{vTv}{v: the contrast vector that defines the parameter of interest}

\item{sigma}{The known noise standard deviation.
If unknown, we recommend a conservative estimate. If it
is left blank, we use the sample variance as a conservative estimate.}

\item{truncation, }{the truncation set (object: intervals)
Computes a confidence interval for the mean of a truncated normal distribution.}

\item{alpha, }{the significance level.}
}
\value{
This function returns a vector of lower and upper confidence limits.
}
\description{
This function computes an equal-tailed (1-alpha) selective confidence intervals.
}
\keyword{internal}
