% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_p_val.R
\name{test_gfl_approx}
\alias{test_gfl_approx}
\title{Monte Carlo significance test for GFLasso, two sided-alternatives only
keywords @internal}
\usage{
test_gfl_approx(
  y,
  Dmat,
  v,
  K_init,
  sigma,
  segment_list,
  ndraws = 1000,
  stop_criteria = "K",
  random_seed = 2021
)
}
\description{
Monte Carlo significance test for GFLasso, two sided-alternatives only
keywords @internal
}
